+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableArray.kt
@@ -347,6 +347,18 @@ public value class ImmutableArray<out T> @PublishedApi internal constructor(
     }
 
     /**
+     * Returns an immutable array containing all the elements expect the first elements that satisfy
+     * the [predicate].
+     */
+    public inline fun dropWhile(predicate: (element: T) -> Boolean): ImmutableArray<T> {
+        var untilIndex = 0
+        while (untilIndex < size && predicate(values[untilIndex])) {
+            untilIndex++
+        }
+        return takeLast(size - untilIndex)
+    }
+
+    /**
      * Returns an immutable array containing only the elements matching the given [predicate].
      */
     public inline fun filter(predicate: (element: T) -> Boolean): ImmutableArray<T> {